factor out db
