--- src/manager/db/mod.rs
@@ -1,5 +1,0 @@
-mod allocations;
-mod claims;
-mod dependent_orders;
-mod lots;
-mod pending_orders;
