+++ src/manager/dependent_orders.rs
@@ -1,12 +1,12 @@
 use super::OrderManager;
+use crate::db;
 use anyhow::{Context, Result};
 use tracing::debug;
 
 impl OrderManager {
     #[tracing::instrument(skip(self, id))]
     pub(super) async fn trigger_dependent_orders(&mut self, id: &str) -> Result<()> {
-        let orders = self
-            .take_dependent_orders(id)
+        let orders = db::take_dependent_orders(&self.db_client, id)
             .await
             .context("Failed to take and delete dependent order")?;
         if !orders.is_empty() {
