+++ src/manager/intents.rs
@@ -1,4 +1,5 @@
 use super::{Claim, OrderManager, Owner, PendingOrder, Position};
+use crate::db;
 use alpaca::{orders::OrderIntent, OrderType, Side};
 use anyhow::{anyhow, Context, Result};
 use chrono::Utc;
@@ -71,14 +72,12 @@ impl OrderManager {
         intent: PositionIntent,
         ticker: &str,
     ) -> Result<()> {
-        let pending_shares = self
-            .get_pending_order_amount_by_ticker(ticker)
+        let pending_shares = db::get_pending_order_amount_by_ticker(&self.db_client, ticker)
             .await
             .context("Failed to get pending order amount")?
             .unwrap_or(0)
             .into();
-        let positions = self
-            .get_positions_by_ticker(&ticker)
+        let positions = db::get_positions_by_ticker(&self.db_client, &ticker)
             .await
             .context("Failed to get positions")?;
         let strategy_shares = positions
@@ -105,18 +104,21 @@ impl OrderManager {
                 Ok(())
             }
             (Some(claim), Some(sent), None) => {
-                self.save_claim(claim)
+                db::save_claim(&self.db_client, claim)
                     .await
                     .context("Failed to save claim")?;
                 let qty = match sent.side {
                     Side::Buy => sent.qty.to_isize().unwrap(),
                     Side::Sell => -(sent.qty.to_isize().unwrap()),
                 };
-                self.save_pending_order(PendingOrder::new(
-                    sent.client_order_id.clone().unwrap(),
-                    ticker.to_string(),
-                    qty as i32,
-                ))
+                db::save_pending_order(
+                    &self.db_client,
+                    PendingOrder::new(
+                        sent.client_order_id.clone().unwrap(),
+                        ticker.to_string(),
+                        qty as i32,
+                    ),
+                )
                 .await
                 .context("Failed to save pending order")?;
                 Ok(self
@@ -126,21 +128,28 @@ impl OrderManager {
             }
             (Some(claim), Some(sent), Some(saved)) => {
                 debug!("Saving dependent order");
-                self.save_dependent_order(&sent.client_order_id.clone().unwrap(), saved)
-                    .await
-                    .context("Failed to save dependent order")?;
-                self.save_claim(claim)
+                db::save_dependent_order(
+                    &self.db_client,
+                    &sent.client_order_id.clone().unwrap(),
+                    saved,
+                )
+                .await
+                .context("Failed to save dependent order")?;
+                db::save_claim(&self.db_client, claim)
                     .await
                     .context("Failed to save claim")?;
                 let qty = match sent.side {
                     Side::Buy => sent.qty.to_isize().unwrap(),
                     Side::Sell => -(sent.qty.to_isize().unwrap()),
                 };
-                self.save_pending_order(PendingOrder::new(
-                    sent.client_order_id.clone().unwrap(),
-                    ticker.to_string(),
-                    qty as i32,
-                ))
+                db::save_pending_order(
+                    &self.db_client,
+                    PendingOrder::new(
+                        sent.client_order_id.clone().unwrap(),
+                        ticker.to_string(),
+                        qty as i32,
+                    ),
+                )
                 .await
                 .context("Failed to save pending order")?;
                 Ok(self
@@ -162,22 +171,19 @@ impl OrderManager {
                     debug!("UpdatePolicy::Retain: No trading needed");
                     return Ok(());
                 }
-                UpdatePolicy::RetainLong => self
-                    .get_positions_by_owner(owner)
+                UpdatePolicy::RetainLong => db::get_positions_by_owner(&self.db_client, owner)
                     .await
                     .context("Failed to get positions")?
                     .into_iter()
                     .filter(|pos| pos.is_short())
                     .collect(),
-                UpdatePolicy::RetainShort => self
-                    .get_positions_by_owner(owner)
+                UpdatePolicy::RetainShort => db::get_positions_by_owner(&self.db_client, owner)
                     .await
                     .context("Failed to get position")?
                     .into_iter()
                     .filter(|pos| pos.is_long())
                     .collect(),
-                UpdatePolicy::Update => self
-                    .get_positions_by_owner(owner)
+                UpdatePolicy::Update => db::get_positions_by_owner(&self.db_client, owner)
                     .await
                     .context("Failed to get positions")?,
             };
@@ -204,14 +210,17 @@ impl OrderManager {
                 position.ticker.clone(),
                 AmountSpec::Shares(-position.shares),
             );
-            self.save_claim(claim)
+            db::save_claim(&self.db_client, claim)
                 .await
                 .context("Failed to save claim")?;
-            self.save_pending_order(PendingOrder::new(
-                order_intent.client_order_id.clone().unwrap(),
-                position.ticker.to_string(),
-                (-position.shares).to_i32().unwrap(),
-            ))
+            db::save_pending_order(
+                &self.db_client,
+                PendingOrder::new(
+                    order_intent.client_order_id.clone().unwrap(),
+                    position.ticker.to_string(),
+                    (-position.shares).to_i32().unwrap(),
+                ),
+            )
             .await
             .context("Failed to save pending order")?;
             Ok(self
