+++ src/manager/order_updates.rs
@@ -1,4 +1,5 @@
 use super::{split_lot, Allocation, Lot, OrderManager};
+use crate::db;
 use alpaca::{Event, OrderEvent, Side};
 use anyhow::{anyhow, Context, Result};
 use chrono::{DateTime, Utc};
@@ -19,13 +20,13 @@ impl OrderManager {
         match event.event {
             Event::Canceled { .. } => {
                 debug!("Order cancelled");
-                self.delete_pending_order_by_id(&event.order.client_order_id)
+                db::delete_pending_order_by_id(&self.db_client, &event.order.client_order_id)
                     .await
                     .context("Failed to delete pending order")?;
             }
             Event::Expired { .. } => {
                 debug!("Order expired");
-                self.delete_pending_order_by_id(&event.order.client_order_id)
+                db::delete_pending_order_by_id(&self.db_client, &event.order.client_order_id)
                     .await
                     .context("Failed to delete pending order")?;
             }
@@ -38,11 +39,11 @@ impl OrderManager {
                     .await
                     .context("Failed to make lot")?;
                 debug!("Deleting pending order");
-                self.delete_pending_order_by_id(&event.order.client_order_id)
+                db::delete_pending_order_by_id(&self.db_client, &event.order.client_order_id)
                     .await
                     .context("Failed to delete pending order")?;
                 debug!("Saving lot");
-                self.save_lot(new_lot.clone())
+                db::save_lot(&self.db_client, new_lot.clone())
                     .await
                     .context("Failed to save lot")?;
                 debug!("Assigning lot");
@@ -58,24 +59,26 @@ impl OrderManager {
                 price, timestamp, ..
             } => {
                 debug!("Partial fill");
-                let pending_order = self
-                    .get_pending_order_by_id(&event.order.client_order_id)
-                    .await
-                    .context("Failed to get pending order")?
-                    .ok_or_else(|| anyhow!("Partial fill received without seeing `new` event"))?;
+                let pending_order =
+                    db::get_pending_order_by_id(&self.db_client, &event.order.client_order_id)
+                        .await
+                        .context("Failed to get pending order")?
+                        .ok_or_else(|| {
+                            anyhow!("Partial fill received without seeing `new` event")
+                        })?;
                 let filled_qty = match event.order.side {
                     Side::Buy => event.order.filled_qty.to_isize().unwrap(),
                     Side::Sell => -(event.order.filled_qty.to_isize().unwrap()),
                 };
                 let pending_qty = pending_order.qty - filled_qty as i32;
-                self.update_pending_order_qty(&id, pending_qty)
+                db::update_pending_order_qty(&self.db_client, &id, pending_qty)
                     .await
                     .context("Failed to update pending order quantity")?;
                 let new_lot = self
                     .make_lot(&id, ticker, timestamp, price, qty)
                     .await
                     .context("Failed to make lot")?;
-                self.save_lot(new_lot.clone())
+                db::save_lot(&self.db_client, new_lot.clone())
                     .await
                     .context("Failed to make lot")?;
                 self.assign_lot(new_lot)
@@ -111,8 +114,7 @@ impl OrderManager {
 
     #[tracing::instrument(skip(self, order_id))]
     async fn previous_fill_data(&self, order_id: &str) -> Result<(Decimal, Decimal)> {
-        let previous_lots = self
-            .get_lots_by_order_id(order_id)
+        let previous_lots = db::get_lots_by_order_id(&self.db_client, order_id)
             .await
             .context("Failed to get lots")?;
         let (prev_qty, prev_price) = previous_lots.iter().fold(
@@ -129,8 +131,7 @@ impl OrderManager {
 
     #[tracing::instrument(skip(self, lot))]
     async fn assign_lot(&mut self, lot: Lot) -> Result<()> {
-        let claims = self
-            .get_claims_by_ticker(&lot.ticker)
+        let claims = db::get_claims_by_ticker(&self.db_client, &lot.ticker)
             .await
             .context("Failed to get claim")?;
         let allocations = split_lot(&claims, &lot);
@@ -138,7 +139,7 @@ impl OrderManager {
             self.adjust_claim(&allocation)
                 .await
                 .context("Failed to adjust claim")?;
-            self.save_allocation(allocation)
+            db::save_allocation(&self.db_client, allocation)
                 .await
                 .context("Failed to save allocation")?;
         }
@@ -148,8 +149,7 @@ impl OrderManager {
     #[tracing::instrument(skip(self, allocation))]
     async fn adjust_claim(&self, allocation: &Allocation) -> Result<()> {
         if let Some(claim_id) = allocation.claim_id {
-            let claim = self
-                .get_claim_by_id(claim_id)
+            let claim = db::get_claim_by_id(&self.db_client, claim_id)
                 .await
                 .context("Failed to get claim")?;
             let amount = match claim.amount {
@@ -163,7 +163,7 @@ impl OrderManager {
                 }
                 _ => unimplemented!(),
             };
-            self.update_claim_amount(claim_id, amount)
+            db::update_claim_amount(&self.db_client, claim_id, amount)
                 .await
                 .context("Failed to update claim amount")?;
         };
